fix(routes-v1): connect donation widget finish page
