+++ routes/admin/authenticated/sidebar/widgets-donation-settings/finish/page.connected.js
@@ -1,0 +1,21 @@
+import { connect } from 'react-redux'
+
+import MobSelectors from '~client/mobrender/redux/selectors'
+import * as MobActions from '~client/mobrender/redux/action-creators'
+
+import Page from './page'
+
+const mapStateToProps = state => {
+  const selectors = MobSelectors(state)
+
+  return {
+    mobilization: selectors.getMobilization(),
+    widget: selectors.getWidget()
+  }
+}
+
+const mapDispatchToProps = {
+  asyncWidgetUpdate: MobActions.asyncUpdateWidget
+}
+
+export default connect(mapStateToProps, mapDispatchToProps)(Page)
