+++ routes/admin/authenticated/sidebar/widgets-donation-settings/finish/page.js
@@ -1,24 +1,8 @@
 import React from 'react'
-import { connect } from 'react-redux'
 import { FormattedMessage } from 'react-intl'
 
-import MobSelectors from '~client/mobrender/redux/selectors'
-import * as MobActions from '~client/mobrender/redux/action-creators'
 import { FormFinishMessage } from '~client/mobilizations/widgets/components'
 import { DonationTellAFriend } from '~client/mobilizations/widgets/__plugins__/donation/components'
-
-const mapStateToProps = state => {
-  const selectors = MobSelectors(state)
-
-  return {
-    mobilization: selectors.getMobilization(),
-    widget: selectors.getWidget()
-  }
-}
-
-const mapDispatchToProps = {
-  asyncWidgetUpdate: MobActions.asyncUpdateWidget
-}
 
 const DonationSettingsFinish = props => (
   <FormFinishMessage
@@ -33,4 +17,4 @@ const DonationSettingsFinish = props => (
   />
 )
 
-export default connect(mapStateToProps, mapDispatchToProps)(DonationSettingsFinish)
+export default DonationSettingsFinish
